commandline programs added
